+++ docs/topics/changelog.md
@@ -3,6 +3,7 @@
 ## 0.4.0-SNAPSHOT
 
 - Update Kotlin to `2.3` and Kotest to `6.1`.
+- The `Expensive` tag object now has internal visibility.
 - Fix issue wherein solutions completing exceptionally would not fail the test.
 - Reading test data now only trims new lines at end of file instead of all whitespace.
   This makes some problems, like [Y2025D06](https://adventofcode.com/2025/day/6) easier to parse.